bug fix:
1.Segment fault when request a non-exist stream.
2.MediaTrack need to stop and re-open when no Session bind to it.

+++ MediaSourceHandler.h
@@ -104,7 +104,7 @@ private:
     bool is_active_;
     std::shared_ptr<std::thread> produce_thread_;
 
-    int read_next_frame();
+    ssize_t read_next_frame();
     void run_producer();
 };
 
@@ -137,7 +137,7 @@ private:
     bool is_active_;
     std::shared_ptr<std::thread> produce_thread_;
 
-    int read_next_frame();
+    ssize_t read_next_frame();
     void run_producer();
 };
